requirements.txt
antspyx==0.3.8
fenics_dolfin==2019.1.0
fenics_ufl==2019.1.0
fslpy>=3.8.2
matplotlib==3.3.2
meshio>=5.3.0
nibabel==3.2.2
numpy==1.21.5
pandas==1.4.1
PyYAML==6.0
scikit_image==0.19.1
scikit_learn==1.0.2
scipy==1.8.0
setuptools==60.9.1
SimpleITK==2.1.0
dcm2niixpy>=1.1.0
torch==1.11.0
pytorch-lightning>=0.8.1
vtk==9.1.0
